JackNewest/glasshouse-capture/requirements.txt
pyserial>=3.5
cobs>=1.2.1
